css fixes and refactor to support additional import options in future
